The HighFieldJob instance no longer requires the job_name as an input.
The HighFieldJob instance no longer requires the structure as an input.

Added a new function to add a layer of adsorbate on stepped surface

Changes to the name of the restart job function.

Added occupancy_smearing as another variable. Changed the order in which the occupancy smearing, convergence_precision and calc_minimize are initialized within the functions.

Changes to set_pbc

Added new functions to restart the jobs based on previous jobs charge density.